Fix bad merge.

+++ core/array_map.go
@@ -214,7 +214,7 @@ func (m *ArrayMap) Merge(other Map) Map {
 	for iter := other.Iter(); iter.HasNext(); {
 		p := iter.Next()
 		res.Set(p.Key, p.Value)
-		if len(res.arr) > HASHMAP_THRESHOLD {
+		if int64(len(res.arr)) > HASHMAP_THRESHOLD {
 			return NewHashMap(m.arr...).Merge(other)
 		}
 	}
